+++ packages/callhierarchy/src/browser/callhierarchy-contribution.ts
@@ -15,17 +15,15 @@
 // *****************************************************************************
 
 import { injectable, inject, postConstruct } from '@theia/core/shared/inversify';
-import { MenuModelRegistry, Command, CommandRegistry, nls } from '@theia/core/lib/common';
+import { MenuModelRegistry, Command, CommandRegistry } from '@theia/core/lib/common';
 import { AbstractViewContribution, OpenViewArguments, KeybindingRegistry } from '@theia/core/lib/browser';
 import { EDITOR_CONTEXT_MENU, CurrentEditorAccess, EditorManager } from '@theia/editor/lib/browser';
 import { CallHierarchyTreeWidget } from './callhierarchy-tree/callhierarchy-tree-widget';
-import { CALLHIERARCHY_ID } from './callhierarchy';
+import { CALLHIERARCHY_ID, CALL_HIERARCHY_LABEL, CALL_HIERARCHY_TOGGLE_COMMAND_ID } from './callhierarchy';
 import { CallHierarchyServiceProvider } from './callhierarchy-service';
 import URI from '@theia/core/lib/common/uri';
 import { ContextKey, ContextKeyService } from '@theia/core/lib/browser/context-key-service';
-
-export const CALL_HIERARCHY_TOGGLE_COMMAND_ID = 'callhierarchy:toggle';
-export const CALL_HIERARCHY_LABEL = nls.localizeByDefault('Call Hierarchy');
+export { CALL_HIERARCHY_LABEL, CALL_HIERARCHY_TOGGLE_COMMAND_ID };
 
 export namespace CallHierarchyCommands {
     export const OPEN = Command.toLocalizedCommand({