+++ packages/process/src/node/process-manager.ts
@@ -14,14 +14,15 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 // *****************************************************************************
 import { injectable, inject, named } from '@theia/core/shared/inversify';
-import { Process } from './process';
 import { Emitter, Event } from '@theia/core/lib/common';
 import { ILogger } from '@theia/core/lib/common/logger';
 import { BackendApplicationContribution } from '@theia/core/lib/node';
+import { ManagedProcessManager, ManagedProcess } from '../common/process-manager-types';
 import { MAX_SAFE_INTEGER } from '@theia/core/lib/common/numbers';
+import { Process } from './process';
 
 @injectable()
-export class ProcessManager implements BackendApplicationContribution {
+export class ProcessManager implements ManagedProcessManager, BackendApplicationContribution {
 
     protected readonly processes: Map<number, Process>;
     protected readonly deleteEmitter: Emitter<number>;
@@ -66,7 +67,7 @@ export class ProcessManager implements BackendApplicationContribution {
      *
      * @param process the process to unregister from this process manager.
      */
-    unregister(process: Process): void {
+    unregister(process: ManagedProcess): void {
         const processLabel = this.getProcessLabel(process);
         this.logger.debug(`Unregistering process. ${processLabel}`);
         if (!process.killed) {
@@ -81,7 +82,7 @@ export class ProcessManager implements BackendApplicationContribution {
         }
     }
 
-    get(id: number): Process | undefined {
+    get(id: number): ManagedProcess | undefined {
         return this.processes.get(id);
     }
 
@@ -99,7 +100,7 @@ export class ProcessManager implements BackendApplicationContribution {
         }
     }
 
-    private getProcessLabel(process: Process): string {
+    private getProcessLabel(process: ManagedProcess): string {
         return `[ID: ${process.id}]`;
     }