+++ packages/task/src/node/task-runner-protocol.ts
@@ -1,5 +1,5 @@
 // *****************************************************************************
-// Copyright (C) 2019 TypeFox and others.
+// Copyright (C) 2022 Ericsson and others.
 //
 // This program and the accompanying materials are made available under the
 // terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,6 +14,20 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 // *****************************************************************************
 
-import { TerminalContribution } from './terminal-widget-impl';
-/** @deprecated @since 1.28.0 import from `terminal-widget-impl` instead. */
-export { TerminalContribution };
+import { TaskConfiguration } from '../common';
+import { Task } from './task';
+
+export const TaskRunner = Symbol('TaskRunner');
+/**
+ * A {@link TaskRunner} knows how to run a task configuration of a particular type.
+ */
+export interface TaskRunner {
+    /**
+     * Runs a task based on the given `TaskConfiguration`.
+     * @param taskConfig the task configuration that should be executed.
+     * @param ctx the execution context.
+     *
+     * @returns a promise of the (currently running) {@link Task}.
+     */
+    run(tskConfig: TaskConfiguration, ctx?: string): Promise<Task>;
+}